+++ packages/core/src/data/ome_zarr/README.md
