+++ packages/core/src/data/ome_zarr/make_zod_types.sh
