+++ packages/core/src/data/ome_zarr_image_source.ts
@@ -1,12 +1,12 @@
 import { Location } from "@zarrita/core";
 import { Readable } from "@zarrita/storage";
 import FetchStore from "@zarrita/storage/fetch";
-import { openArray, openGroup } from "./zarrita/open";
-import WebFileSystemStore from "./zarrita/web_file_system_store";
+import { openArray, openGroup } from "./zarr/open";
+import WebFileSystemStore from "./zarr/web_file_system_store";
 import { OmeZarrImageLoader } from "../data/ome_zarr_image_loader";
 import {
   omeZarrToZarrVersion,
-  parseOmeNgffImage,
+  parseOmeZarrImage,
 } from "./ome_zarr_hcs_metadata_loader";
 
 /** Opens an OME-Zarr multiscale image Zarr group from either a URL or local directory. */
@@ -33,7 +33,7 @@ export class OmeZarrImageSource {
 
   public async open(): Promise<OmeZarrImageLoader> {
     const root = await openGroup(this.location);
-    const adaptedOmeImage = parseOmeNgffImage(root.attrs);
+    const adaptedOmeImage = parseOmeZarrImage(root.attrs);
     const omeVersion = adaptedOmeImage.originalVersion;
     const images = adaptedOmeImage.multiscales;
     if (images.length !== 1) {
