+++ packages/core/src/data/zarr/open.ts
