+++ packages/core/src/index.ts
@@ -24,7 +24,7 @@ export { Box2 } from "./math/box2";
 export { Box3 } from "./math/box3";
 
 export type { Region } from "./data/region";
-export type { Image as OmeNgffImage } from "./data/ome_ngff/0.4/image";
+export type { Image as OmeZarrImage } from "./data/ome_zarr/0.4/image";
 export {
   loadOmeroChannels,
   loadOmeroDefaultZ,
